Temporarily undo the rich text split fix instead of the paragraph hook
The paragraph hook only handles Enter on an empty paragraph inside a
container. The plain split lives in rich text's Enter listener.
