+++ packages/block-editor/src/components/rich-text/event-listeners/enter.js
@@ -41,10 +41,10 @@ export default ( props ) => ( element ) => {
 		} else if ( onSplitAtEnd && start === end && end === text.length ) {
 			event.preventDefault();
 			onSplitAtEnd();
+		} else if ( onReplace && onSplit ) {
+			event.__deprecatedOnSplit = true;
 		} else if (
 			! supportsSplitting &&
-			// The deprecated onSplit is flagged on the beforeinput event.
-			! ( onReplace && onSplit ) &&
 			! disableLineBreaks &&
 			! event.defaultPrevented
 		) {
@@ -70,22 +70,8 @@ export default ( props ) => ( element ) => {
 		}
 	}
 
-	function onBeforeInput( event ) {
-		if ( event.inputType !== 'insertParagraph' ) {
-			return;
-		}
-		const { onReplace, onSplit } = props.current;
-		if ( onReplace && onSplit ) {
-			event.__deprecatedOnSplit = true;
-		}
-	}
-
-	function onDefaultBeforeInput( event ) {
-		if (
-			event.defaultPrevented ||
-			( event.inputType !== 'insertParagraph' &&
-				event.inputType !== 'insertLineBreak' )
-		) {
+	function onDefaultKeyDown( event ) {
+		if ( event.defaultPrevented ) {
 			return;
 		}
 
@@ -96,6 +82,12 @@ export default ( props ) => ( element ) => {
 			return;
 		}
 
+		if ( event.keyCode !== ENTER ) {
+			return;
+		}
+
+		// On ENTER, we ALWAYS want to prevent the default browser behaviour
+		// at this last interception point.
 		event.preventDefault();
 	}
 
@@ -103,10 +95,10 @@ export default ( props ) => ( element ) => {
 
 	// Attach the listener to the window so parent elements have the chance to
 	// prevent the default behavior.
-	const unsubscribeDefaultBeforeInput = subscribeDelegatedListener(
+	const unsubscribeDefaultKeyDown = subscribeDelegatedListener(
 		defaultView,
-		'beforeinput',
-		onDefaultBeforeInput
+		'keydown',
+		onDefaultKeyDown
 	);
 	// Capture phase so this runs before ancestor (writing flow) bubble
 	// handlers, matching the timing of the previous raw element listener.
@@ -116,15 +108,8 @@ export default ( props ) => ( element ) => {
 		onKeyDown,
 		true
 	);
-	const unsubscribeBeforeInput = subscribeOwnedListener(
-		element,
-		'beforeinput',
-		onBeforeInput,
-		true
-	);
 	return () => {
-		unsubscribeDefaultBeforeInput();
+		unsubscribeDefaultKeyDown();
 		unsubscribeKeyDown();
-		unsubscribeBeforeInput();
 	};
 };
